CEDS-1958 Rename forms.Mapping to fix compilation warnings
